[Python API] Rename offline_transformations in the new api
