+++ src/bindings/python/src/pyopenvino/core/offline_transformations.cpp
@@ -37,7 +37,7 @@ namespace py = pybind11;
 void regmodule_offline_transformations(py::module m) {
     // TODO: change the submodule name according to the description in 69196
     py::module m_offline_transformations =
-        m.def_submodule("offline_transformations_pybind", "Offline transformations module");
+        m.def_submodule("offline_transformations", "Offline transformations module");
 
     m_offline_transformations.def(
         "apply_moc_transformations",
