+++ src/bindings/python/tests/test_transformations/test_offline_api.py
@@ -4,7 +4,7 @@
 import os
 import numpy as np
 # TODO: change the module name according to the description in 69196
-from openvino.offline_transformations_pybind import apply_moc_transformations, apply_pot_transformations, \
+from openvino.offline_transformations import apply_moc_transformations, apply_pot_transformations, \
     apply_low_latency_transformation, apply_pruning_transformation, apply_make_stateful_transformation, \
     compress_model_transformation, serialize
