+++ tools/mo/openvino/tools/mo/back/offline_transformations.py
@@ -9,9 +9,9 @@ from openvino.tools.mo.utils.cli_parser import parse_transform
 
 def get_available_transformations():
     try:
-        from openvino.offline_transformations_pybind import apply_low_latency_transformation # pylint: disable=import-error,no-name-in-module
-        from openvino.offline_transformations_pybind import apply_make_stateful_transformation # pylint: disable=import-error,no-name-in-module
-        from openvino.offline_transformations_pybind import apply_pruning_transformation # pylint: disable=import-error,no-name-in-module
+        from openvino.offline_transformations import apply_low_latency_transformation # pylint: disable=import-error,no-name-in-module
+        from openvino.offline_transformations import apply_make_stateful_transformation # pylint: disable=import-error,no-name-in-module
+        from openvino.offline_transformations import apply_pruning_transformation # pylint: disable=import-error,no-name-in-module
         return {
             'MakeStateful': apply_make_stateful_transformation,
             'LowLatency2': apply_low_latency_transformation,
@@ -33,11 +33,11 @@ def apply_user_transformations(func: object, transforms: list):
 
 
 def apply_moc_transformations(func: object):
-    from openvino.offline_transformations_pybind import apply_moc_transformations  # pylint: disable=import-error,no-name-in-module
+    from openvino.offline_transformations import apply_moc_transformations  # pylint: disable=import-error,no-name-in-module
     apply_moc_transformations(func, False)
 
 def compress_model(func: object):
-    from openvino.offline_transformations_pybind import compress_model_transformation  # pylint: disable=import-error,no-name-in-module
+    from openvino.offline_transformations import compress_model_transformation  # pylint: disable=import-error,no-name-in-module
     compress_model_transformation(func)
 
 
@@ -46,7 +46,7 @@ def apply_offline_transformations(input_model: str, argv: argparse.Namespace):
     # to produce correct mapping
     extract_names = argv.framework in ['tf', 'mxnet', 'kaldi']
 
-    from openvino.offline_transformations_pybind import generate_mapping_file, serialize  # pylint: disable=import-error,no-name-in-module
+    from openvino.offline_transformations import generate_mapping_file, serialize  # pylint: disable=import-error,no-name-in-module
     from openvino.frontend import FrontEndManager, FrontEnd  # pylint: disable=no-name-in-module,import-error
     from openvino.tools.mo.back.preprocessing import apply_preprocessing  # pylint: disable=no-name-in-module,import-error
