+++ tools/mo/openvino/tools/mo/moc_frontend/serialize.py
@@ -28,7 +28,7 @@ def moc_emit_ir(ngraph_function: Model, argv: argparse.Namespace):
 
     orig_model_name = os.path.normpath(os.path.join(output_dir, argv.model_name))
 
-    from openvino.offline_transformations_pybind import serialize # pylint: disable=import-error,no-name-in-module
+    from openvino.offline_transformations import serialize # pylint: disable=import-error,no-name-in-module
     serialize(ngraph_function, (orig_model_name + ".xml").encode('utf-8'), (orig_model_name + ".bin").encode('utf-8'))
 
     del argv.feManager
